move game 1 to another file for better structure
